+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Flip to true to test it. To see if it is working, try processing some files whil
 When enabled you should see delivery to many different topics, one for each core and worker that is running.
 With it off, all messages will be delivered to the dgs topic. When on, DeltaFiles will be cached locally and made eventually
 consistent in the database. This decreases processing latency but does not give a realtime view of the state in the UI.
+- Added `/blackhole` endpoint to `deltafi-egress-sink`.  The endpoint will always return a 200 and never write content to disk, acting as a noop egress destination.  The endpoint will take a latency parameter to add latency to the post response (i.e. `/blackhole?latency=0.1` to add 100ms latency)
 
 ### Changed
 
@@ -19,6 +20,7 @@ consistent in the database. This decreases processing latency but does not give 
 ### Removed
 
 ### Fixed
+- A metric bug with illegal characters in tags
 - DeltaFiles in the JOINING stage are now considered "in-flight" when getting DeltaFile stats.
 
 ### Tech-Debt/Refactor